Proxy grpc from wg-sock-stats to json
This seems odd/ circuitous, but it will allow us to run our maps app in a container, passing the socket from the above _to_ that container.

This in turn means our app can run non-root, nice and separate from wireguard and anything which requires root. It also means that building other clients to load data from wg-sock-stats into other places (log databases and so on) can continue to make use of the benefits we receive from using grpc and protobufs